requirements.txt
pydantic
fastapi
uvicorn
sqlacodegen
SQLAlchemy
mysql-connector-python
GeoAlchemy2
pymysql
bcrypt
passlib
python-multipart
PyJWT


